+++ tests/testthat/test-colwise-filter.R
@@ -26,12 +26,12 @@ test_that("aborts on empty selection", {
 test_that("aborts when supplied funs() or list", {
   expect_error(
     filter_all(mtcars, list(~. > 0)),
-    "`.vars_predicate` must be a call to `all_vars()` or `any_vars()`, not a list",
+    "`.vars_predicate` must be a function or a call to `all_vars()` or `any_vars()`, not a list",
     fixed = TRUE
   )
   expect_error(
     filter_all(mtcars, funs(. > 0)),
-    "`.vars_predicate` must be a call to `all_vars()` or `any_vars()`, not a `fun_list` object",
+    "`.vars_predicate` must be a function or a call to `all_vars()` or `any_vars()`, not a `fun_list` object",
     fixed = TRUE
   )
 })
@@ -55,4 +55,14 @@ test_that("filter_if and filter_all includes grouping variables (#3351, #3480)",
 
   res <- filter_if(tbl, is.integer, all_vars(. > 1))
   expect_true(all(res$gr1 > 1))
+})
+
+test_that("can supply functions to scoped filters", {
+  exp <- as.list(mtcars[c(8, 9, 21), ])
+
+  out <- mtcars %>% filter_at(c("cyl", "am"), ~ .x == 4 | .x == 0)
+  expect_identical(as.list(out), exp)
+
+  out <- mtcars %>% filter_at(c("cyl", "am"), function(.x) .x == 4 | .x == 0)
+  expect_identical(as.list(out), exp)
 })